requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pandas
google-generativeai
